refactor(resources): move skill and bundle routes
